==> 4 - get_relevant_orders.ipynb <==
import pandas as pd
import spacy
import pickle
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

from collections import defaultdict
from itertools import chain

# Load data
processed_queries = pd.read_csv('data\\processed_queries.csv')
processed_orders = pd.read_csv('data\\processed_orders.csv')

processed_queries.head()

processed_orders.head()

TF-IDF Similarity

# Load the TF-IDF matrix and vectoriser
with open("models\\tfidf\\tfidf_matrix.pkl", "rb") as matrix_file:
    tfidf_matrix = pickle.load(matrix_file)

with open("models\\tfidf\\tfidf_vectorizer.pkl", "rb") as matrix_file:
    tfidf_vec = pickle.load(matrix_file)

# Get scores for all queries loaded
tfidf_scores = []

for idx, row in processed_queries.iterrows():
    query_tfidf_vector = tfidf_vec.transform([row['query']])
    similarities = cosine_similarity(query_tfidf_vector, tfidf_matrix)

    tfidf_scores.append(similarities[0])

Named entity similarity

# Load named entity bag of words
with open("models\\ner_bow\\ner_bow_model.pkl", "rb") as bow_file:
    named_entity_vec = pickle.load(bow_file)

with open("models\\ner_bow\\ner_bow_matrix.pkl", "rb") as matrix_file:
    named_entity_matrix = pickle.load(matrix_file)

# Get scores for all queries loaded
ner_scores_temp = []

for idx, row in processed_queries.iterrows():
    query_ner_vector = named_entity_vec.transform([row['named_entities']])

    similarities = cosine_similarity(named_entity_matrix, query_ner_vector)

    ner_scores_temp.append(similarities)

ner_scores = []

for scores in ner_scores_temp:
    ner_scores.append(list(chain(*scores)))

Noun phrase similarity

# Load named entity bag of words
with open("models\\nph_bow\\nph_bow_model.pkl", "rb") as bow_file:
    noun_phrases_vec = pickle.load(bow_file)

with open("models\\nph_bow\\nph_bow_matrix.pkl", "rb") as matrix_file:
    noun_phrases_matrix = pickle.load(matrix_file)

# Get scores for all queries loaded
nph_scores_temp = []

for idx, row in processed_queries.iterrows():
    query_nph_vector = noun_phrases_vec.transform([row['chunks']])

    similarities = cosine_similarity(noun_phrases_matrix, query_nph_vector)

    nph_scores_temp.append(similarities)

nph_scores = []

for scores in nph_scores_temp:
    nph_scores.append(list(chain(*scores)))

Overall document similarity

# Load the spaCy model
nlp = spacy.load("en_core_web_lg")

ovr_scores = []

for idx, row in processed_queries.iterrows():
    doc1 = nlp(row['summary'])

    query_scores = []

    for idx, row in processed_orders.iterrows():
        doc2 = nlp(row['summary'])

        # Calculate similarity
        similarity_score = doc1.similarity(doc2)

        query_scores.append(similarity_score)

    ovr_scores.append(query_scores)

Create overall scores

import pandas as pd

# Initialize an empty DataFrame with specific columns
columns = ['tfidf_scores','named_entity_scores', 'noun_phrase_scores', 'doc_scores']
similarity_scores = pd.DataFrame(columns={col: [] for col in columns})

similarity_scores['tfidf_scores'] = tfidf_scores
similarity_scores['named_entity_scores'] = ner_scores
similarity_scores['noun_phrase_scores'] = nph_scores
similarity_scores['doc_scores'] = ovr_scores


# Calculate the average similarity score for each document for each query
average_list = [sum(x) / len(x) for x in zip(tfidf_scores, ner_scores, nph_scores, ovr_scores)]

# Convert average scores to indexed dictionary
averages_scores = []

for item in average_list:
    averages_dict = defaultdict(list)
    for idx, score in enumerate(item):
        averages_dict[idx] = score
    
    averages_scores.append(averages_dict)

# Get the top 3 scoring documents for each query
top_3_list = []

for idx, scores in enumerate(averages_scores):
    sorted_averages = dict(sorted(averages_scores[idx].items(), key=lambda item: item[1], reverse=True))
    top_3 = list(sorted_averages.items())[:3]
    top_3_list.append(top_3)

similarity_scores['top'] = top_3_list

similarity_scores
